requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
dj-database-url==3.0.1
Django==5.2.5
django-sendgrid-v5==1.3.0
gunicorn==23.0.0
idna==3.10
mailtrap==2.1.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
python-dotenv==1.1.1
python-http-client==3.3.7
sendgrid==6.12.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.9.0
